Precision is a four-rung ladder, and the creator audit found three more holes
FP64 is the new bottom rung: double width, nobody trains like that any
more, and it is what a company is left with before it has researched
anything. FP32, BF16 and FP8 are each bought. The 1.0 anchor moves down to
FP64 for the third time and the same reason — whatever a player gets for
free has to be the reference, or the economy is retuned underneath them.
FP8 stays at 3.36, because the top of the ladder is an anchor too and a
fourth rung is not a reason to move the ceiling.

You asked whether the locks could be bugged past. They could, and the audit
that says so is now a test file that never touches the panel:

The projection did not enforce the parameter ceiling or the precision gate
at all. Both lived only in TryStartTraining, so the creator priced a run
happily and GO refused it, with no warning attached to anything the player
had touched. Both are checked in Project now.

And the training calendar was still wrong, in the other direction from last
time. The creator quotes compute time plus the safety stage; the run
counted only petaflop-days, so a model priced at eleven days finished in
one and every week the SAFETY stage charged for was handed straight back.
The run has two clocks now, like the upgrade programmes always have.

The floor is the safety stage alone rather than the whole quoted duration.
Pinning a run to its original estimate would mean buying accelerators
mid-run bought nothing, and a test that has guarded that promise for months
said so within a minute of my trying it.

Saves go to v34. Every existing company is granted single precision and
trains at exactly the width it always did.

660 EditMode, 6 PlayMode, scene 107.

+++ Assets/_ScalingLaws/Scripts/Simulation/ModelBlueprint.cs
@@ -37,7 +37,7 @@ namespace ScalingLaws.Simulation
             DatasetSource dataSources,
             ModelType type = ModelType.General,
             string family = null,
-            TrainingPrecision precision = TrainingPrecision.Float32,
+            TrainingPrecision precision = TrainingPrecision.Float64,
             ModelShape shape = ModelShape.Balanced,
             DeduplicationPass deduplication = DeduplicationPass.Standard,
             int cutoffMonthsBack = 0,
